pyproject.toml
[project]
name = "urlbox-screenshot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["python-dotenv>=1.1.0", "urlbox>=1.0.6"]
